[ticket/10124] Remove `$_POST` dependency from acp_permissions
The `set_permission` method is dependent upon the $_POST array which
makes it impossible to call that method with the data that has to be
set. This commit introduces a new method parameter which can be used to
pass the information to the method.

The old referances to the `$_POST` array and `request_var` calls are
still in place for backward compatibility but could be removed in the
develop branch.

PHPBB3-10124